# Optimization: move select_relateds to creation of ProjectResource