this removes an expensive query on /project/ views for directories

/projects/firefox/browser/

NEW: 6 queries run, total 0.071 seconds
OLD: 7 queries run, total 0.139 seconds

/projects/firefox/browser/chrome/browser/safebrowsing/

NEW: 6 queries run, total 0.052 seconds
OLD: 7 queries run, total 0.096 seconds